Create make_pakna.py
Makes a list of all possible test strings (over the alphabet of the ur-language pakna). After creating the test file, one runs PhoMEnt with a grammar file. PhoMEnt produces a 'score file' for every string in the test file, which is used as the base distribution for generating a corpus.